Modernize packaging & testing
