+++ examples/test_bed.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-from __future__ import print_function
+from __future__ import annotations
 
 import argparse, collections, os, subprocess, sys, time
