+++ pytest.ini
@@ -1,0 +1,16 @@
+[pytest]
+
+addopts =
+    --cov=smartcrop
+    --cov-fail-under=50
+    --cov-report term-missing
+    --cov-report html
+    --cov-report xml
+    --doctest-modules
+    --ignore=examples
+    --pylint
+    # --flake8
+
+doctest_optionflags =
+    ELLIPSIS
+    NORMALIZE_WHITESPACE
