+++ smartcrop/__init__.py
@@ -1,0 +1,3 @@
+from .library import SmartCrop, saturation, thirds
+
+__all__ = ['SmartCrop', 'saturation', 'thirds']
